requirements.txt
flask
flask_script
flask_restful
flask_login
flask_mail
flask_jwt_extended
flask_sqlalchemy
flask_migrate
flask_redis
click
gevent
gunicorn
pymysql
xlwt
pycryptodome
rsa
redis
celery
git+https://github.com/cangmean/fet.git